refactor(*): Move infra to pkg and add close for sourcepool
