+++ cmd/kuiper/main.go
@@ -18,8 +18,8 @@ import (
 	"bufio"
 	"fmt"
 	"github.com/lf-edge/ekuiper/internal/conf"
-	"github.com/lf-edge/ekuiper/internal/infra"
 	"github.com/lf-edge/ekuiper/internal/pkg/model"
+	"github.com/lf-edge/ekuiper/pkg/infra"
 	"github.com/urfave/cli"
 	"io/ioutil"
 	"net/rpc"
