+++ internal/topo/checkpoint/responder.go
@@ -16,7 +16,7 @@ package checkpoint
 
 import (
 	"fmt"
-	"github.com/lf-edge/ekuiper/internal/infra"
+	"github.com/lf-edge/ekuiper/pkg/infra"
 )
 
 type Responder interface {
