+++ internal/topo/node/watermark.go
@@ -17,10 +17,10 @@ package node
 import (
 	"context"
 	"fmt"
-	"github.com/lf-edge/ekuiper/internal/infra"
 	"github.com/lf-edge/ekuiper/internal/xsql"
 	"github.com/lf-edge/ekuiper/pkg/api"
 	"github.com/lf-edge/ekuiper/pkg/ast"
+	"github.com/lf-edge/ekuiper/pkg/infra"
 	"math"
 	"sort"
 )
